chore: remove old newgroupCardAvatars component
